requirements.txt
numpy
paddlepaddle
loguru
tabulate
